updated sys decommission report
git-svn-id: svn://107.170.10.188:8080/Semoss/trunk@2786 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

+++ src/prerna/algorithm/impl/specific/tap/SysDecommissionScheduleOptimizer.java
@@ -341,7 +341,7 @@ public class SysDecommissionScheduleOptimizer implements IAlgorithm{
 			Object[] rowForSys = new Object[maxYears+1];
 			rowForSys[0] = systemList.get(sysInd);
 			for(int i=1;i<=maxYears;i++)
-				rowForSys[i] = systemSiteMatrix.get(i-1)[sysInd];
+				rowForSys[i] = (systemSiteMatrix.get(i-1)[sysInd]).intValue();
 			list.add(rowForSys);
 		}
